fix differentiated calculator

+++ Models/CreditInfo.cs
@@ -11,7 +11,7 @@ namespace CreditCalculator.Models
         public decimal TotalSum { get; set; }
         public DateTime DateOfIssue { get; set; }
         public DateTime DateOfClosing { get; set; }
-        public float Percent { get; set; }
+        public decimal Percent { get; set; }
         public CalculationType CreditType { get; set; }
         public int PaymentDay { get; set; }
     }
